+++ lib/rrj/janus/messages/admin.rb
@@ -15,7 +15,7 @@ module RubyRabbitmqJanus
 
         # Return options to message for rabbitmq
         def options
-          properties.options_admin(@type)
+          properties.options_admin(type)
         rescue
           raise Errors::Janus::MessageAdmin::Options
         end