+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [2.1.0] - 2017-
+## [2.1.0] - 2017-0r65-13
 ### [Added]
 - Janus Instance management
 - Initializer for rake task